Add support multiple config for token providers
- This is a change to the existing functionality but is backwards compatible and hence feature parity.
- Code before this change was supporting just one configuration for an ID provider. So, if an org had multiple azure configurations which needed supporting, that was not possible. The usecase could be to have different azure configuration for say prod and non-prod.
- The configurations could now be provided as a list against the ID provider. The targets can then be added within that configuration rather than a global configuration for all the targets which was the case before.

+++ e2e/e2e_suite_test.go
@@ -133,6 +133,9 @@ func resetDefaults() {
 }
 
 func cleanup() {
+	if ospreyconfig == nil {
+		return
+	}
 	if err := os.Remove(ospreyconfig.Kubeconfig); err != nil {
 		Expect(os.IsNotExist(err)).To(BeTrue())
 	}
